+++ packages/go_router/lib/src/delegate.dart
@@ -50,30 +50,8 @@ class GoRouterDelegate extends RouterDelegate<RouteMatchList>
   RouteMatchList _matches = RouteMatchList.empty();
   final Map<String, int> _pushCounts = <String, int>{};
 
-  /// Pushes the given location onto the page stack
-  void push(RouteMatch match) {
-    // Remap the pageKey to allow any number of the same page on the stack
-    final String fullPath = match.fullpath;
-    final int count = (_pushCounts[fullPath] ?? 0) + 1;
-    _pushCounts[fullPath] = count;
-    final ValueKey<String> pageKey = ValueKey<String>('$fullPath-p$count');
-    final RouteMatch newPageKeyMatch = RouteMatch(
-      route: match.route,
-      subloc: match.subloc,
-      fullpath: match.fullpath,
-      encodedParams: match.encodedParams,
-      queryParams: match.queryParams,
-      extra: match.extra,
-      error: match.error,
-      pageKey: pageKey,
-    );
-
-    _matches.push(newPageKeyMatch);
-    notifyListeners();
-  }
-
-  /// Pushes the given location asynchronously onto the page stack with a promise.
-  Future<T?> pushAsync<T extends Object?>(RouteMatch match) {
+  /// Pushes the given location onto the page stack with an optional promise.
+  Future<T?> push<T extends Object?>(RouteMatch match) {
     // Remap the pageKey to allow any number of the same page on the stack.
     final String fullPath = match.fullpath;