+++ packages/go_router/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 4.3.0
 
-- Adds `pushAsync` and `pushNamedAsync` to `go_router` making it posible to await results from a `pop`.
+- Adds a promise option to `push` and `pushNamed` making it posible to await results from a `pop`.
 - Adds an optional parameter to the `pop` method, allowing to return values from a `pop`.
 
 ## 4.2.7